+++ lib/shelly/helpers.rb
@@ -191,11 +191,13 @@ More info at http://git-scm.com/book/en/Git-Basics-Getting-a-Git-Repository}
 
     def apps_table(apps)
       apps.map do |app|
-        msg = if app.state == "deploy_failed" || app.state == "configuration_failed"
-          " (deployment log: `shelly deploys show last -c #{app.code_name}`)"
-        end
+        msg = info_last_deploy_logs(app) if app.in_deploy_failed_state?
         [app.code_name, "|  #{app.state_description}#{msg}"]
       end
+    end
+
+    def info_last_deploy_logs(app)
+      " (deployment log: `shelly deploys show last -c #{app}`)"
     end
   end
 end